third_party/blink/web_tests/external/wpt/css/css-fonts/calc-in-font-variation-settings-expected.txt

This is a testharness.js-based test.
[FAIL] calc() specified value serialization in font-variation-settings
  assert_equals: expected "\\"XHGT\\" calc(0.7)" but got ""
Harness: the test ran to completion.
